Avoid warning of compilation.

+++ eugene/src/Parametrization/CrossingOver.cc
@@ -77,10 +77,10 @@ void CrossingOver::Crosseval (void)
       ChildrenPopulation[1]->data->P = algo->Population[n2]->data->P;
       Crossover(ChildrenPopulation[0], ChildrenPopulation[1]);
       algo->Par = ChildrenPopulation[0]->data->P;
-      for (k=0; k<algo->Para.size(); k++) algo->Para[k] = algo->Par[algo->ParaPar[k]];
+      for (k=0; k<(int)algo->Para.size(); k++) algo->Para[k] = algo->Par[algo->ParaPar[k]];
       ChildrenPopulation[0]->RawFitness = OPTIM.ParaEvaluate();
       algo->Par = ChildrenPopulation[1]->data->P;
-      for (k=0; k<algo->Para.size(); k++) algo->Para[k] = algo->Par[algo->ParaPar[k]];
+      for (k=0; k<(int)algo->Para.size(); k++) algo->Para[k] = algo->Par[algo->ParaPar[k]];
       ChildrenPopulation[1]->RawFitness = OPTIM.ParaEvaluate();
       /* Here everything as been calculated, we just have to crossover */
       algo->SA->SATournament4(ChildrenPopulation[0], ChildrenPopulation[1],
