+++ eugene/src/Parametrization/Mutating.cc
@@ -75,7 +75,7 @@ void Mutating::Muteval (void)
       ChildrenPopulation[0]->data->P = algo->Population[j]->data->P;
       Mutation(ChildrenPopulation[0]) ;
       algo->Par = ChildrenPopulation[0]->data->P;
-      for (k=0; k<algo->Para.size(); k++) algo->Para[k] = algo->Par[algo->ParaPar[k]];
+      for (k=0; k<(int)algo->Para.size(); k++) algo->Para[k] = algo->Par[algo->ParaPar[k]];
       ChildrenPopulation[0]->RawFitness = OPTIM.ParaEvaluate();
       /* Here everything as been calculated, we just have to crossover */
       algo->SA->SATournament2(ChildrenPopulation[0], algo->Population[j]);
